Add own implementation
